# beam-web/openapi.json
{
  "openapi": "3.2.0",
  "info": {
    "title": "Beam Server API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/media/{id}": {
      "get": {
        "tags": [
          "media"
        ],
        "summary": "Fetch a single media item's full metadata by id.",
        "operationId": "getMediaDetail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Media id (movie or show UUID).",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Origin",
            "in": "header",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "Referer",
            "in": "header",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Problem"
                    },
                    {
                      "properties": {
                        "type": {
                          "type": "string",
                          "const": "about:blank"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Forbidden; Cross-origin request rejected; Malformed Origin/Referer header",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request; Invalid media id",
            "content": {
              "application/problem+json": {
                "schema": {
                  "oneOf": [
                    {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/Problem"
                        },
                        {
                          "properties": {
                            "type": {
                              "type": "string",
                              "const": "about:blank"
                            }
                          }
                        }
                      ]
                    },
                    {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/Problem"
                        },
                        {
                          "properties": {
                            "type": {
                              "type": "string",
                              "const": "https://beam.justinchung.net/reference/errors/#invalid-media-id"
                            }
                          }
                        }
                      ]
                    }
                  ]
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaMetadata"
                }
              }
            }
          },
          "404": {
            "description": "Media not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Problem"
                    },
                    {
                      "properties": {
                        "type": {
                          "type": "string",
                          "const": "https://beam.justinchung.net/reference/errors/#media-not-found"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Problem"
                    },
                    {
                      "properties": {
                        "type": {
                          "type": "string",
                          "const": "https://beam.justinchung.net/reference/errors/#internal"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "BeamSession": []
          }
        ]
      }
    },
    "/v1/media/{id}/sources": {
      "get": {
        "tags": [
          "media"
        ],
        "summary": "List the playable/downloadable source files for a playable media id.",
        "description": "Accepts a movie id or an episode id (both are \"playable\" ids). A show id is\nrejected with 400 -- shows have no files of their own, so callers request\nsources for the show's individual episode ids instead. An episode with no\nfiles yet returns an empty array (a valid, \"not yet playable\" response).",
        "operationId": "getMediaSources",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Media id (movie or show UUID).",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Origin",
            "in": "header",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "Referer",
            "in": "header",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Problem"
                    },
                    {
                      "properties": {
                        "type": {
                          "type": "string",
                          "const": "about:blank"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Forbidden; Cross-origin request rejected; Malformed Origin/Referer header",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request; Invalid media id; Sources are not available at the show level",
            "content": {
              "application/problem+json": {
                "schema": {
                  "oneOf": [
                    {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/Problem"
                        },
                        {
                          "properties": {
                            "type": {
                              "type": "string",
                              "const": "about:blank"
                            }
                          }
                        }
                      ]
                    },
                    {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/Problem"
                        },
                        {
                          "properties": {
                            "type": {
                              "type": "string",
                              "const": "https://beam.justinchung.net/reference/errors/#invalid-media-id"
                            }
                          }
                        }
                      ],
                      "title": "Invalid media id"
                    },
                    {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/Problem"
                        },
                        {
                          "properties": {
                            "type": {
                              "type": "string",
                              "const": "https://beam.justinchung.net/reference/errors/#sources-not-available-for-show"
                            }
                          }
                        }
                      ],
                      "title": "Sources are not available at the show level"
                    }
                  ]
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MediaSource"
                  },
                  "type": "array"
                }
              }
            }
          },
          "404": {
            "description": "Media not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Problem"
                    },
                    {
                      "properties": {
                        "type": {
                          "type": "string",
                          "const": "https://beam.justinchung.net/reference/errors/#media-not-found"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Problem"
                    },
                    {
                      "properties": {
                        "type": {
                          "type": "string",
                          "const": "https://beam.justinchung.net/reference/errors/#internal"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "BeamSession": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Problem": {
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "maximum": 65535.0,
            "minimum": 0.0,
            "format": "uint16"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "type",
          "status"
        ],
        "title": "Problem Details",
        "description": "An RFC 9457 problem detail."
      },
...
      "ShowMetadata": {
        "properties": {
          "id": {
            "type": "string",
            "description": "Stable identifier for this show (UUID)."
          },
          "title": {
            "$ref": "#/components/schemas/Title",
            "description": "Title of the show"
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional description of the show"
          },
          "year": {
            "type": [
              "integer",
              "null"
            ],
            "maximum": 4294967295.0,
            "minimum": 0.0,
            "format": "uint32",
            "description": "Year the show was released"
          },
          "poster_url": {
            "type": [
              "string",
              "null"
            ],
            "description": "Path of the show's own poster image, served by the artwork route;\nabsent when the show has no poster."
          },
          "backdrop_url": {
            "type": [
              "string",
              "null"
            ],
            "description": "Path of the show's own backdrop image, served by the artwork route;\nabsent when the show has no backdrop."
          },
          "seasons": {
            "items": {
              "$ref": "#/components/schemas/SeasonMetadata"
            },
            "type": "array",
            "description": "List of seasons in the show"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "seasons"
        ]
      },
...
      "MediaMetadata": {
        "oneOf": [
          {
            "properties": {
              "Show": {
                "$ref": "#/components/schemas/ShowMetadata"
              }
            },
            "additionalProperties": false,
            "type": "object",
            "required": [
              "Show"
            ]
          },
          {
            "properties": {
              "Movie": {
                "$ref": "#/components/schemas/MovieMetadata"
              }
            },
            "additionalProperties": false,
            "type": "object",
            "required": [
              "Movie"
            ]
          }
        ],
        "description": "Media metadata"
      },
      "VideoSourceInfo": {
        "properties": {
          "codec": {
            "type": "string"
          },
          "width": {
            "type": "integer",
            "maximum": 4294967295.0,
            "minimum": 0.0,
            "format": "uint32"
          },
          "height": {
            "type": "integer",
            "maximum": 4294967295.0,
            "minimum": 0.0,
            "format": "uint32"
          },
          "bit_rate": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0.0,
            "format": "uint64"
          },
          "hdr_format": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object",
        "required": [
          "codec",
          "width",
          "height"
        ]
      },
      "AudioSourceInfo": {
        "properties": {
          "codec": {
            "type": "string"
          },
          "language": {
            "type": [
              "string",
              "null"
            ]
          },
          "channels": {
            "type": "integer",
            "maximum": 65535.0,
            "minimum": 0.0,
            "format": "uint16"
          },
          "is_default": {
            "type": "boolean"
          }
        },
        "type": "object",
        "required": [
          "codec",
          "channels",
          "is_default"
        ]
      },
      "MediaSource": {
        "properties": {
          "file_id": {
            "type": "string"
          },
          "size_bytes": {
            "type": "integer",
            "minimum": 0.0,
            "format": "uint64"
          },
          "mime_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "container_format": {
            "type": [
              "string",
              "null"
            ]
          },
          "duration_secs": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "video": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VideoSourceInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "audio_tracks": {
            "items": {
              "$ref": "#/components/schemas/AudioSourceInfo"
            },
            "type": "array"
          },
          "stream_url": {
            "type": "string",
            "description": "Direct-play stream URL for this file (Range-request capable, inline)."
          },
          "download_url": {
            "type": "string",
            "description": "Download URL for this file (Range-request capable, `Content-Disposition: attachment`)."
          }
        },
        "type": "object",
        "required": [
          "file_id",
          "size_bytes",
          "audio_tracks",
          "stream_url",
          "download_url"
        ],
        "description": "One playable/downloadable version of a media item. Beam models multiple\ndeliverable qualities/editions as distinct source files rather than\ntranscoding on demand (see ADR-0004: never live-transcode); this is how a\nclient picks among them for constrained-bandwidth playback."
      },
...
    },
    "securitySchemes": {
      "BeamSession": {
        "type": "apiKey",
        "name": "beam_session",
        "in": "cookie",
        "description": "Opaque session cookie issued by the OIDC login flow (ADR-0003). Hashed at rest; never recoverable from the API."
      }
    }
  },
...
}